04/12/2023

+++ src/app/medicamentos/medicamentos.component.ts
@@ -2,6 +2,8 @@ import { Component, Input, OnChanges, OnInit, SimpleChanges, inject } from '@ang
 import { MedicamentosService } from "../medicamentos.service";
 import { MedicamentosDocument } from "../medicamentos-document";
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { YesNOComponent } from '../yes-no/yes-no.component';
+import { MatDialog } from '@angular/material/dialog';
 @Component({
   selector: 'app-medicamentos',
   templateUrl: './medicamentos.component.html',
@@ -11,6 +13,7 @@ export class MedicamentosComponent implements OnInit,OnChanges{
   @Input() uuid: string = '';
   medicamentosService = inject(MedicamentosService);
   matSnackBar = inject(MatSnackBar);
+  medicamentosDialog= inject(MatDialog);
   medicamentos: MedicamentosDocument[] = [];
   
   ngOnInit(): void {
@@ -27,14 +30,16 @@ export class MedicamentosComponent implements OnInit,OnChanges{
     })  
   }
 
-  remove(id:string){
-    console.log('entre remove');
-    this.medicamentosService.deleteMedicamento(id).subscribe({next: document=>{
-      this.medicamentosService.getMedicamentos(this.uuid).subscribe((medicamentos: MedicamentosDocument[]) => {
-        this.medicamentos = medicamentos;
-        this.matSnackBar.open(`Medicamento borrado ${document.name}`,'',{duration:500});
-      })
-    }})
-    
+  remove(id:string,name:string){
+    this.medicamentosDialog.open(YesNOComponent, { role:'alertdialog',enterAnimationDuration:500,disableClose: true, data: { action: 'Borrar', description: `Desea borrar al medicamento ${name}` } }).afterClosed().subscribe((respuesta: boolean) => {
+      if (respuesta === true) {
+        this.medicamentosService.deleteMedicamento(id).subscribe({next: document=>{
+          this.medicamentosService.getMedicamentos(this.uuid).subscribe((medicamentos: MedicamentosDocument[]) => {
+            this.medicamentos = medicamentos;
+            this.matSnackBar.open(`Medicamento borrado ${document.name}`,'',{duration:500});
+          });
+        }});
+      }
+    });
   }
 }
